Refactor: 콘텐츠 페이지 url 폴더구조 수정
