+++ src/app/content/faq/[id]/layout.tsx
