+++ src/app/content/weekInfo/[id]/layout.tsx
@@ -1,0 +1,14 @@
+import AnimatedNavBar from '<prefix>/components/common/bar/navbar/animatedNavBar';
+
+type ContentProps = {
+  children: React.ReactNode;
+};
+
+export default function FAQLayout({ children }: ContentProps) {
+  return (
+    <>
+      {children}
+      <AnimatedNavBar />
+    </>
+  );
+}
