+++ src/components/content/faq/faqItem.tsx
@@ -12,7 +12,7 @@ type FaqItemProps = {
 export default function FaqItem({ faqItem }: FaqItemProps) {
   const { id, question, image, views } = faqItem;
   return (
-    <Link className='flex h-81 w-full gap-12' href={`/faq/${id}`}>
+    <Link className='flex h-81 w-full gap-12' href={`/content/faq/${id}`}>
       {image && (
         <Image
           src={image as string}
