+++ src/components/content/weekInfo/weekInfoItem.tsx
@@ -1,7 +1,3 @@
-import Image from 'next/image';
-import React from 'react';
-import EyeIcon from '/public/svgs/light/icon-eye.svg';
-import { formatNumberWithCommas } from '<prefix>/shared/utils/format';
 import Link from 'next/link';
 import { WeekInfoResponse } from '<prefix>/shared/types/content';
 
@@ -9,10 +5,10 @@ type FaqItemProps = {
   weekInfoItem: WeekInfoResponse;
 };
 
-export default function WeekInfoItem({ weekInfoItem }: FaqItemProps) {
+export default function ({ weekInfoItem }: FaqItemProps) {
   const { id, week, summary, maternity, fetus } = weekInfoItem;
   return (
-    <Link className='flex h-81 w-full gap-12' href={`/weekInfo/${id}`}>
+    <Link className='flex h-81 w-full gap-12' href={`/content/weekInfo/${id}`}>
       <div className='flex flex-col gap-4'>
         <p className='text-overflow shrink-0 text-body-04 text-neutral-900'>
           <span className='text-primary'>{week}주차 </span>
